+++ delphi-frontend/src/test/resources/au/com/integradev/delphi/grammar/AttributesUnit.pas
@@ -1,0 +1,32 @@
+unit AttributesUnit;
+
+interface
+
+[xyz]
+[assembly: xyz]
+
+type
+  [xyz]
+  [assembly: xyz]
+  [xyz]
+  TMyType = Integer;
+
+[xyz]
+[assembly: xyz]
+
+implementation
+
+[xyz]
+procedure Foo;
+[xyz]
+var
+  [xyz]
+  Bar: Integer;
+begin
+
+end;
+
+[xyz]
+[assembly: xyz]
+
+end.